Split window_specification_internal out of window_specification
